[FEAT] 설정에 유저 정보 표시
